+++ models/modelsMaterials.js
@@ -5,9 +5,9 @@ const modelsMaterials = async (queryAsync) => {
     try {
         const modelsMaterialsTable = await queryAsync(
             `CREATE TABLE IF NOT EXISTS models_materials(
-                id INT AUTO_INCREMENT PRIMARY KEY,
-                model_id INT,
-                material_id INT,
+                id BIGINT(64) AUTO_INCREMENT PRIMARY KEY,
+                model_id BIGINT(64),
+                material_id BIGINT(64),
                 created_at TIMESTAMP,
                 FOREIGN KEY (model_id) REFERENCES models(id),
                 FOREIGN KEY (material_id) REFERENCES materials(id)
